PRL-6550: use same bearer token for multiple tests (#127)
PRL-6550: use same bearer token for multiple tests
