+++ e2e/common/idamLoginHelper.ts
@@ -1,7 +1,7 @@
 import Config from "../config.ts";
 import { Page } from "@playwright/test";
 import { UserCredentials, UserLoginInfo } from "./types.ts";
-import { setupUser } from "./createCitizenUser/createCitizenUser.ts";
+import { setupUser } from "./idamCreateCitizenUserApiHelper.ts";
 
 export class IdamLoginHelper {
   private static fields: UserLoginInfo = {
@@ -26,14 +26,11 @@ export class IdamLoginHelper {
         `#skiplinktarget:text("Sign in or create an account")`,
       );
     }
-
     const userCredentials: UserCredentials = Config.getUserCredentials(user);
-
     if (userCredentials) {
       await page.fill(this.fields.username, userCredentials.email);
       await page.fill(this.fields.password, userCredentials.password);
       await page.click(this.submitButton);
-
       await page
         .context()
         .storageState({ path: Config.sessionStoragePath + `${user}.json` });
@@ -46,7 +43,12 @@ export class IdamLoginHelper {
     page: Page,
     application: string,
   ): Promise<void> {
-    const userInfo = await setupUser(); // Create a new citizen user
+    const token = process.env.BEARER_TOKEN;
+    if (!token) {
+      console.error("Bearer token is not defined in the environment variables");
+      return;
+    }
+    const userInfo = await setupUser(token);
     if (!userInfo) {
       console.error("Failed to set up citizen user");
       return;
