+++ e2e/tests/citizen/caseDashboard/caseDashboard.spec.ts
@@ -5,7 +5,6 @@ import IdamLoginHelper from "../../../common/idamLoginHelper";
 
 test.describe("Manage citizen cases case dashboard tests. @citizenFrontend @crossbrowserCitizenFrontend", (): void => {
   test.beforeEach(async ({ page }) => {
-    // Sign in as a citizen user before each test
     await IdamLoginHelper.signInCitizenUser(
       page,
       Config.citizenFrontendBaseURL,
