+++ e2e/tests/citizen/createCase/C100/c100ThirdRowMiroJourney.spec.ts
@@ -5,7 +5,6 @@ import IdamLoginHelper from "../../../../common/idamLoginHelper";
 
 test.describe("C100 Citizen Application tests on the third MIRO set. @citizenFrontend @crossbrowserCitizenFrontend", (): void => {
   test.beforeEach(async ({ page }) => {
-    // Sign in as a citizen user before each test
     await IdamLoginHelper.signInCitizenUser(
       page,
       Config.citizenFrontendBaseURL,
@@ -468,10 +467,9 @@ test.describe("C100 Citizen Application tests on the third MIRO set. @citizenFro
 
 test.describe("C100 Citizen Application accessibility tests on the third row journey set. @accessibilityCitizenFrontend", (): void => {
   test.beforeEach(async ({ page }) => {
-    // Sign in as a citizen user before each test
     await IdamLoginHelper.signInCitizenUser(
-        page,
-        Config.citizenFrontendBaseURL,
+      page,
+      Config.citizenFrontendBaseURL,
     );
   });
   test(`C100 Citizen Application with the following options:
