+++ e2e/tests/citizen/createCase/C100/c100YouNeedMIAMDocumentSigned.spec.ts
@@ -5,7 +5,6 @@ import IdamLoginHelper from "../../../../common/idamLoginHelper";
 
 test.describe("Create Citizen Application but you don't have a MIAM document. @citizenFrontend @crossbrowserCitizenFrontend", (): void => {
   test.beforeEach(async ({ page }) => {
-    // Sign in as a citizen user before each test
     await IdamLoginHelper.signInCitizenUser(
       page,
       Config.citizenFrontendBaseURL,
