+++ package.json
@@ -9,7 +9,6 @@
     "@axe-core/playwright": "^4.10.0",
     "dotenv": "^16.4.5",
     "playwright-core": "^1.46.1",
-    "qs": "^6.13.0",
     "typescript": "^5.5.4",
     "uuid": "^11.0.2"
   },
@@ -19,7 +18,6 @@
     "@playwright/test": "^1.46.1",
     "@types/jest": "^29.5.12",
     "@types/node": "^22.4.1",
-    "@types/qs": "^6",
     "axios": "^1.7.4",
     "jest": "^29.7.0",
     "pre-commit": "^1.2.2",
